refactor: moved the interfaces
