+++ aiogram_nats/core/interfaces/interfaces/message.py
@@ -1,6 +1,19 @@
 from typing import Protocol
 
-from aiogram_nats.core.entities.message import MessageSendScheduled
+from aiogram_nats.core.entities.message import MessageDeletionScheduled, MessageSendScheduled
+
+
+class MessageRemover(Protocol):
+
+    """
+    Class represents a message remover.
+
+    It provides a method to remove a message from the system.
+    """
+
+    async def remove(self, message: MessageDeletionScheduled) -> None:
+        """Removes a message from the system."""
+        raise NotImplementedError
 
 
 class MessageSender(Protocol):
